feat: move utils into a package
