+++ packages/core/src/types.ts
@@ -1,11 +1,3 @@
-export type Variants = "outlined" | "filled" | "full";
-
-export interface IconTree {
-  tag: string;
-  attr: { [key: string]: string };
-  child: IconTree[];
-}
-
 export interface IconBaseProps extends React.SVGAttributes<SVGElement> {
   className?: string;
   title?: string;
