+++ packages/utils/CHANGELOG.md
@@ -1,0 +1,48 @@
+## [0.2.5](https://github.com/rocketclimb/rocketicons/compare/v0.3.3...v0.3.4) (2024-05-25)
+
+### Bug Fixes
+
+- **icons** improve class name handling on native ([6575100f](https://github.com/rocketclimb/rocketicons/commit/6575100fa60dd11dd3086225d25ebc532805cda2))
+
+## [0.2.4](https://github.com/rocketclimb/rocketicons/compare/v0.3.1...v0.3.2) (2024-05-21)
+
+### Bug Fixes
+
+- attrToString function and increase test coverage ([41bdbf54](https://github.com/rocketclimb/rocketicons/commit/41bdbf54d302017a14dac796e5bf02dbe7bd2d27))
+- **icons** allow multiple icon classes to coexist ([8883e3e5](https://github.com/rocketclimb/rocketicons/commit/8883e3e5eda09af88be332383d0afd6cc7cbc13e))
+
+## [0.2.3](https://github.com/rocketclimb/rocketicons/compare/v0.2.4-release...v0.3.0) (2024-05-17)
+
+## [0.2.2](https://github.com/rocketclimb/rocketicons/compare/v0.2.3-release...v0.2.4) (2024-05-09)
+
+- No significant changes
+
+## [0.2.1](https://github.com/rocketclimb/rocketicons/compare/v0.2.1...v0.2.2) (2024-05-05)
+
+### Bug Fixes
+
+- **esm module** tailwind plugin and core package ([11fa38a](https://github.com/rocketclimb/rocketicons/commit/11fa38a9385a937f69afc499148b9063728e4594))
+
+## [0.2.0](https://github.com/rocketclimb/rocketicons/compare/v0.1.0...v0.2.0) (2024-04-30)
+
+### Features
+
+- **icons** initial version set ([22fab59](https://github.com/rocketclimb/rocketicons/commit/22fab597d3528ee244f97d0a2a94eacab1c9dc20))
+- **icons** fixing branch changes ([d392dfb](https://github.com/rocketclimb/rocketicons/commit/d392dfbce5cfd53ccd3ae8819aecae40d3d77886))
+
+### Documentation
+
+- changelog title format ([d66ca4d](https://github.com/rocketclimb/rocketicons/commit/d66ca4d60c9d595f75a65352d3aba432a9e3b173))
+
+### Continuous integration
+
+- **releaser** bump packages versions and update changelog for release/liftoff ([45f1f70](https://github.com/rocketclimb/rocketicons/commit/45f1f70b81a84c35147a69065bcfc476decc7a2d))
+
+## [0.1.0](https://github.com/rocketclimb/rocketicons/compare/v0.0.0...v0.1.0) (2024-04-28)
+
+### Features
+
+- icons as React component
+- icons as React Native component
+- Tailwind classes handling
+- CommonJS and ESM
