+++ packages/utils/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@rocketicons/utils",
+  "version": "0.2.5",
+  "license": "MIT",
+  "main": "./dist/index.js",
+  "files": [
+    "dist/"
+  ],
+  "scripts": {
+    "build": "tsc -b ./tsconfig.cjs.json ./tsconfig.esm.json ./tsconfig.types.json && tsconfig-replace-paths",
+    "postbuild": "rocketclimb-sh cp \"dist-esm/**/*.js\" \"dist/**/*.mjs\" && rocketclimb-sh fix-esm ./dist",
+    "clean": "rocketclimb-sh rm -fR dist dist-esm && rocketclimb-sh rm -fR *.tsbuildinfo",
+    "clean:o": "rocketclimb-sh rm -fR node_modules",
+    "format": "prettier --check . --write --ignore-path ../../.gitignore",
+    "lint": "eslint src/**/*",
+    "rebuild": "npm run clean && npm run build",
+    "test": "jest ./src",
+    "typecheck": "tsc --noEmit"
+  }
+}
