+++ packages/utils/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./tree-2-string";
+export * from "./types";
 export { styleToString } from "./style-to-string";
 export { mergeStyles } from "./merge-styles";
 export { handleClassName } from "./handle-class-name";
