+++ packages/utils/src/types.ts
@@ -1,0 +1,7 @@
+export type Variants = "outlined" | "filled" | "full";
+
+export interface IconTree {
+  tag: string;
+  attr: { [key: string]: string };
+  child: IconTree[];
+}
