+++ packages/utils/tsconfig.esm.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "outDir": "./dist-esm",
+    "module": "esnext"
+  }
+}
